# packages/contracts/openapi/control-api.yaml
openapi: 3.1.0
info:
  title: Pal Control API
  version: 0.4.0
  license:
    name: MIT
    identifier: MIT
  description: >-
    Browser-facing API for Palworld operations. The official Palworld REST API
    and native mod bridge are internal adapters and must not be exposed directly.
    Management operations require a configured administrator API key and are
    additionally restricted to the loopback management boundary. Privileged
    economy and season writes also require a TOTP code and an auditable reason.
    Player-portal sessions use a separate cookie identity and never authorize
    management operations. Non-loopback deployments require HTTPS at a trusted
    reverse proxy in addition to these application controls.
servers:
  - url: http://127.0.0.1:5180/api/v1
    description: Local development
...
paths:
...
  /admin/economy-continuity/capacity:
    get:
      tags: [ExtractionAdmin]
      summary: Estimate authoritative-state backup capacity without deleting data
      operationId: getEconomyContinuityCapacity
      responses:
        "200":
          description: Current bytes, retained estimate, required free space and capacity decision
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: "#/components/responses/ExtractionError"
  /admin/economy-continuity/retention-plan:
    get:
      tags: [ExtractionAdmin]
      summary: List old economy backups eligible for separately approved cleanup
      operationId: planEconomyBackupRetention
      responses:
        "200":
          description: Plan-only result; this endpoint never deletes a backup or world
          content:
            application/json:
              schema:
                type: object
                required: [deleteAutomatically, candidates]
                properties:
                  deleteAutomatically:
                    type: boolean
                    const: false
                  candidates:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
        default:
          $ref: "#/components/responses/ExtractionError"
  /admin/economy-continuity/snapshots:
    post:
      tags: [ExtractionAdmin]
      summary: Create a WAL-consistent economy snapshot under maintenance
      operationId: createEconomySnapshot
      parameters:
        - $ref: "#/components/parameters/AdminTotp"
        - $ref: "#/components/parameters/AdminReason"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [serverId, worldId]
              properties:
                serverId:
                  type: string
                  minLength: 1
                  maxLength: 64
                worldId:
                  type: string
                  pattern: "^[A-Fa-f0-9]{32}$"
                idempotencyKey:
                  type: [string, "null"]
                  minLength: 16
                  maxLength: 128
                  description: Use the durable rollover economy-backup step key.
      responses:
        "200":
          description: Verified snapshot manifest; an idempotent replay returns the same backupId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EconomySnapshotManifest"
        default:
          $ref: "#/components/responses/ExtractionError"
  /admin/economy-continuity/snapshots/{serverId}/{backupId}/verify:
    get:
      tags: [ExtractionAdmin]
      summary: Verify snapshot manifest, files and SQLite integrity
      operationId: verifyEconomySnapshot
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
        - name: backupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verified snapshot manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EconomySnapshotManifest"
        default:
          $ref: "#/components/responses/ExtractionError"
  /admin/economy-continuity/snapshots/{serverId}/{backupId}/stage:
    post:
      tags: [ExtractionAdmin]
      summary: Restore one economy snapshot to an isolated fail-closed staging directory
      description: >-
        A replay for the same backup and expected world revalidates and returns
        the already atomically published staging result. It never creates a
        second staging copy or promotes staging into production.
      operationId: stageEconomySnapshot
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
        - name: backupId
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/AdminTotp"
        - $ref: "#/components/parameters/AdminReason"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expectedWorldId]
              properties:
                expectedWorldId:
                  type: string
                  pattern: "^[A-Fa-f0-9]{32}$"
      responses:
        "200":
          description: Hash, integrity, replay, active-world, ledger and fail-closed result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: "#/components/responses/ExtractionError"
  /admin/economy-continuity/snapshots/{serverId}/{backupId}/post-snapshot:
    get:
      tags: [ExtractionAdmin]
      summary: List events and unresolved transactions requiring recovery reconciliation
      operationId: listPostEconomySnapshotTransactions
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
        - name: backupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reconciliation items after the snapshot sequence plus current blockers
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/EconomyPendingTransaction"
        default:
          $ref: "#/components/responses/ExtractionError"
  /admin/weekly-rollover/operations:
    post:
      tags: [ExtractionAdmin]
      summary: Prepare or replay a deterministic persistent weekly rollover
      operationId: prepareWeeklyRollover
      parameters:
        - $ref: "#/components/parameters/AdminTotp"
        - $ref: "#/components/parameters/AdminReason"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [serverId, fromSeasonId, fromWorldId, targetWorldId, rulesVersion]
              properties:
                serverId:
                  type: string
                  minLength: 1
                  maxLength: 64
                fromSeasonId:
                  type: string
                  format: uuid
                fromWorldId:
                  type: string
                  pattern: "^[A-Fa-f0-9]{32}$"
                targetWorldId:
                  type: string
                  pattern: "^[A-Fa-f0-9]{32}$"
                rulesVersion:
                  type: string
                  minLength: 1
                  maxLength: 128
      responses:
        "200":
          description: Persistent operation and deterministic key for the required phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyRolloverResponse"
        default:
          $ref: "#/components/responses/ExtractionError"
  /admin/weekly-rollover/operations/active:
    get:
      tags: [ExtractionAdmin]
      summary: Resume the server's one incomplete weekly rollover after restart
      operationId: getActiveWeeklyRollover
      parameters:
        - name: serverId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incomplete rollover and its required step key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyRolloverResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/weekly-rollover/operations/{operationId}:
    get:
      tags: [ExtractionAdmin]
      summary: Get one persistent weekly rollover
      operationId: getWeeklyRollover
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Rollover state and its required step key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyRolloverResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/season-settlement-jobs/voucher-expiry:
    post:
      tags: [ExtractionAdmin]
      summary: Freeze the one versioned SeasonVoucher-expiry job for a season
      operationId: prepareSeasonVoucherExpiry
      parameters:
        - $ref: "#/components/parameters/AdminTotp"
        - $ref: "#/components/parameters/AdminReason"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [seasonId, rulesVersion]
              properties:
                seasonId:
                  type: string
                  format: uuid
                rulesVersion:
                  type: string
                  minLength: 1
                  maxLength: 128
      responses:
        "200":
          description: Frozen per-account expiry job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeasonSettlementJob"
        default:
          $ref: "#/components/responses/ExtractionError"
...
components:
  securitySchemes:
    adminApiKey:
      type: apiKey
      in: header
      name: X-Pal-Admin-Key
      description: >-
        Administrator API key. Configure only a SHA-256 digest at rest; never
        store the plaintext key in repository configuration or logs.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    playerSessionCookie:
      type: apiKey
      in: cookie
      name: pal_player_session
      description: >-
        Opaque player session set only after successful in-game OTP verification.
        The deployment uses an HttpOnly, SameSite=Strict cookie scoped to
        /api/v1/player and enables Secure behind HTTPS. Authenticated player
        identity is derived exclusively from this cookie, never from a userId
        query parameter or business request body.
  parameters:
    AdminTotp:
      name: X-Pal-Admin-Totp
      in: header
      required: true
      description: Six-digit RFC 6238 TOTP code required for privileged writes.
      schema:
        type: string
        pattern: "^[0-9]{6}$"
...
  responses:
...
    ExtractionError:
      description: Extraction-mode validation, state, persistence, or adapter error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
security:
  - adminApiKey: []